Do not use HRTimeSpan on korgw/korge and remove confusing hr*Seconds Extension properties for the sake of simplicity (#36)

+++ samples/s3d/src/commonMain/kotlin/utils.kt
@@ -15,7 +15,7 @@ private suspend fun Stage3D.orbit(v: View3D, distance: Double, time: TimeSpan) {
 		val angle = 360.degrees * ratio
 		camera.positionLookingAt(
 			cos(angle) * distance, 0.0, sin(angle) * distance, // Orbiting camera
-			v.transform.translation.x, v.transform.translation.y, v.transform.translation.z
+			v.transform.translation.x.toDouble(), v.transform.translation.y.toDouble(), v.transform.translation.z.toDouble()
 		)
 	}
 }
